+++ CHANGELOG.md
@@ -14,6 +14,9 @@
         - PluginController
         - NetworkManager
         - BLEManager
+        - DFUFeature
+        - DetectorTemperatureFeature
+        - StripChartsFeature
 - 2026-08-19 4.2.9
     - only show ENLIGHTEN version in window title in Expert mode (simplifies manual screenshots)
     - DALAI