+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - LibraryMatching
         - renamed assets/example_data/matching_library to default_matching_library
         - renamed EnlightenSpectra/matching_library to user_library_spectra
+        - defer to plugins on table contention
     - added WhatsThis
         - LibraryMatching
         - EtalonCorrection