+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 4.2 XS V2 Commercial Release
 
-- 2026-??-?? 4.2.10
+- 2026-08-21 4.2.10
     - persist CorrectionStatus visibility
     - add headers when copying StripChart data to clipboard
     - LibraryMatching